refactoring: dropdown component  - component  - docs
