+++ docs/components/dropdown.md
@@ -1,7 +1,7 @@
 <script setup>
-import DropdownPlacementExample from './dropdown/examples/DropdownPlacementExample.vue';
-import DropdownListGroupExample from './dropdown/examples/DropdownListGroupExample.vue';
-import DropdownTriggerExample from './dropdown/examples/DropdownTriggerExample.vue';
+import FwbDropdownExamplePlacement from './dropdown/examples/FwbDropdownExamplePlacement.vue'
+import FwbDropdownExampleListGroup from './dropdown/examples/FwbDropdownExampleListGroup.vue'
+import FwbDropdownExampleTrigger from './dropdown/examples/FwbDropdownExampleTrigger.vue'
 </script>
 
 # Vue Dropdown - Flowbite
@@ -16,122 +16,85 @@ Original reference: [https://flowbite.com/docs/components/dropdowns/](https://fl
 
 The dropdown component can be used to show a list of menu items when clicking on an element such as a button and hiding it when focusing outside of the triggering element.
 
+## Dropdown - placement
 
-## Props - placement
-
+<fwb-dropdown-example-placement />
 ```vue
-<script setup>
-import { Dropdown } from 'flowbite-vue'
-</script>
 <template>
-  <dropdown placement="bottom" text="Bottom">
-    Any content here
-  </dropdown>
-  <dropdown placement="top">
-    <template #trigger="{ toggle }">
-      <Button @click="toggle">
-        Top
-        <template #suffix>
-          <svg class="ml-2 w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"></path></svg>
-        </template>
-      </Button>
-    </template>
-    <div class="p-2">
-      Padding content
-    </div>
-  </dropdown>
-  <dropdown placement="right" text="Right">
-    <Spinner size="6" class="m-4" />
-  </dropdown>
-  <dropdown placement="left" text="Left">
-    hello world
-  </dropdown>
+  <fwb-dropdown text="Bottom">
+    <p class="p-2">Dropdown content here</p>
+  </fwb-dropdown>
+  <fwb-dropdown placement="top" text="Top">
+    <p class="p-2">Dropdown content here</p>
+  </fwb-dropdown>
+  <fwb-dropdown placement="right" text="Right">
+    <p class="p-2">Dropdown content here</p>
+  </fwb-dropdown>
+  <fwb-dropdown placement="left" text="Left">
+    <p class="p-2">Dropdown content here</p>
+  </fwb-dropdown>
 </template>
+
+<script setup>
+import { FwbDropdown } from 'flowbite-vue'
+</script>
 ```
 
-<DropdownPlacementExample />
+## Dropdown with List Group
 
-
-## List Group
-
+<fwb-dropdown-example-list-group />
 ```vue
-<script setup>
-import { Dropdown, ListGroup, ListGroupItem } from 'flowbite-vue'
-</script>
 <template>
-  <dropdown text="Bottom">
+  <fwb-dropdown text="Menu">
     <list-group>
       <list-group-item>
-        <template #prefix>
-          <svg class="w-4 h-4 fill-current" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" d="M18 10a8 8 0 11-16 0 8 8 0 0116 0zm-6-3a2 2 0 11-4 0 2 2 0 014 0zm-2 4a5 5 0 00-4.546 2.916A5.986 5.986 0 0010 16a5.986 5.986 0 004.546-2.084A5 5 0 0010 11z" clip-rule="evenodd"></path></svg>
-        </template>
         Profile
       </list-group-item>
       <list-group-item>
-        <template #prefix>
-          <svg class="w-4 h-4 fill-current" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path d="M5 4a1 1 0 00-2 0v7.268a2 2 0 000 3.464V16a1 1 0 102 0v-1.268a2 2 0 000-3.464V4zM11 4a1 1 0 10-2 0v1.268a2 2 0 000 3.464V16a1 1 0 102 0V8.732a2 2 0 000-3.464V4zM16 3a1 1 0 011 1v7.268a2 2 0 010 3.464V16a1 1 0 11-2 0v-1.268a2 2 0 010-3.464V4a1 1 0 011-1z"></path></svg>
-        </template>
         Settings
       </list-group-item>
       <list-group-item>
-        <template #prefix>
-          <svg class="w-4 h-4 fill-current" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" d="M5 3a2 2 0 00-2 2v10a2 2 0 002 2h10a2 2 0 002-2V5a2 2 0 00-2-2H5zm0 2h10v7h-2l-1 2H8l-1-2H5V5z" clip-rule="evenodd"></path></svg>
-        </template>
         Messages
       </list-group-item>
       <list-group-item>
-        <template #prefix>
-          <svg class="w-4 h-4 fill-current" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" d="M2 9.5A3.5 3.5 0 005.5 13H9v2.586l-1.293-1.293a1 1 0 00-1.414 1.414l3 3a1 1 0 001.414 0l3-3a1 1 0 00-1.414-1.414L11 15.586V13h2.5a4.5 4.5 0 10-.616-8.958 4.002 4.002 0 10-7.753 1.977A3.5 3.5 0 002 9.5zm9 3.5H9V8a1 1 0 012 0v5z" clip-rule="evenodd"></path></svg>
-        </template>
         Download
       </list-group-item>
     </list-group>
-  </dropdown>
+  </fwb-dropdown>
 </template>
+
+<script setup>
+import { FwbDropdown, ListGroup, ListGroupItem } from 'flowbite-vue'
+</script>
 ```
 
-<DropdownListGroupExample />
+## Dropdown - trigger
 
-## Slot - trigger
-
+<fwb-dropdown-example-trigger />
 ```vue
-<script setup>
-import { Dropdown, ListGroup, ListGroupItem } from 'flowbite-vue'
-</script>
 <template>
-  <dropdown text="Bottom">
+  <fwb-dropdown text="Bottom">
     <template #trigger>
-      <span>Click trigger</span>
+      <span>Custom Trigger Element</span>
     </template>
     <list-group>
       <list-group-item>
-        <template #prefix>
-          <svg class="w-4 h-4 fill-current" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" d="M18 10a8 8 0 11-16 0 8 8 0 0116 0zm-6-3a2 2 0 11-4 0 2 2 0 014 0zm-2 4a5 5 0 00-4.546 2.916A5.986 5.986 0 0010 16a5.986 5.986 0 004.546-2.084A5 5 0 0010 11z" clip-rule="evenodd"></path></svg>
-        </template>
         Profile
       </list-group-item>
       <list-group-item>
-        <template #prefix>
-          <svg class="w-4 h-4 fill-current" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path d="M5 4a1 1 0 00-2 0v7.268a2 2 0 000 3.464V16a1 1 0 102 0v-1.268a2 2 0 000-3.464V4zM11 4a1 1 0 10-2 0v1.268a2 2 0 000 3.464V16a1 1 0 102 0V8.732a2 2 0 000-3.464V4zM16 3a1 1 0 011 1v7.268a2 2 0 010 3.464V16a1 1 0 11-2 0v-1.268a2 2 0 010-3.464V4a1 1 0 011-1z"></path></svg>
-        </template>
         Settings
       </list-group-item>
       <list-group-item>
-        <template #prefix>
-          <svg class="w-4 h-4 fill-current" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" d="M5 3a2 2 0 00-2 2v10a2 2 0 002 2h10a2 2 0 002-2V5a2 2 0 00-2-2H5zm0 2h10v7h-2l-1 2H8l-1-2H5V5z" clip-rule="evenodd"></path></svg>
-        </template>
         Messages
       </list-group-item>
       <list-group-item>
-        <template #prefix>
-          <svg class="w-4 h-4 fill-current" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" d="M2 9.5A3.5 3.5 0 005.5 13H9v2.586l-1.293-1.293a1 1 0 00-1.414 1.414l3 3a1 1 0 001.414 0l3-3a1 1 0 00-1.414-1.414L11 15.586V13h2.5a4.5 4.5 0 10-.616-8.958 4.002 4.002 0 10-7.753 1.977A3.5 3.5 0 002 9.5zm9 3.5H9V8a1 1 0 012 0v5z" clip-rule="evenodd"></path></svg>
-        </template>
         Download
       </list-group-item>
     </list-group>
-  </dropdown>
+  </fwb-dropdown>
 </template>
+
+<script setup>
+import { FwbDropdown, ListGroup, ListGroupItem } from 'flowbite-vue'
+</script>
 ```
-
-
-<DropdownTriggerExample />
