--- src/components/Dropdown/types.ts
@@ -1,1 +1,0 @@
-export type DropdownPlacement = 'top' | 'bottom' | 'left' | 'right'
