+++ src/components/FwbDropdown/types.ts
@@ -1,0 +1,1 @@
+export type DropdownPlacement = 'top' | 'bottom' | 'left' | 'right'
