+++ src/index.ts
@@ -13,11 +13,11 @@ export { default as FwbButton } from './components/FwbButton/FwbButton.vue'
 export { default as FwbButtonGroup } from './components/FwbButtonGroup/FwbButtonGroup.vue'
 export { default as FwbCard } from './components/FwbCard/FwbCard.vue'
 export { default as FwbCarousel } from './components/FwbCarousel/FwbCarousel.vue'
+export { default as FwbDropdown } from './components/FwbDropdown/FwbDropdown.vue'
 
 export { default as Spinner } from './components/Spinner/Spinner.vue'
 export { default as Tabs } from './components/Tabs/Tabs.vue'
 export { default as Tab } from './components/Tabs/components/Tab/Tab.vue'
-export { default as Dropdown } from './components/Dropdown/Dropdown.vue'
 export { default as FlowbiteThemable } from './components/utils/FlowbiteThemable/FlowbiteThemable.vue'
 export { default as FlowbiteThemableChild } from './components/utils/FlowbiteThemable/components/FlowbiteThemableChild/FlowbiteThemableChild.vue'
 export { default as Footer } from './components/Footer/Footer.vue'
